# laravel/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'MRC API'
  description: 'REST API учебного портала (LMS). Роли: студент, преподаватель, администратор.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8085'
paths:
...
  /api/profile/email:
    post:
      summary: 'Смена email'
      operationId: Email
      description: 'Сохраняет новый адрес в `new_email` и отправляет письмо с подтверждением. Текущий `email` не меняется, пока пользователь не перейдёт по ссылке из письма.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Письмо с подтверждением отправлено на новый адрес.'
                  user:
                    id: 2
                    role: teacher
                    role_label: Преподаватель
                    first_name: Иван
                    second_name: Сергеевич
                    last_name: Петров
                    full_name: 'Петров Иван Сергеевич'
                    email: teacher@institute.local
                    new_email: new.teacher@institute.local
                    email_verified_at: null
                    phone: null
                    is_active: true
                    default_path: /api/gradebooks
                    gradebooks_count: 2
                    created_at: '2026-05-16T12:00:00+00:00'
                    updated_at: '2026-05-16T12:15:00+00:00'
                properties:
                  message:
                    type: string
                    example: 'Письмо с подтверждением отправлено на новый адрес.'
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 2
                      role:
                        type: string
                        example: teacher
                      role_label:
                        type: string
                        example: Преподаватель
                      first_name:
                        type: string
                        example: Иван
                      second_name:
                        type: string
                        example: Сергеевич
                      last_name:
                        type: string
                        example: Петров
                      full_name:
                        type: string
                        example: 'Петров Иван Сергеевич'
                      email:
                        type: string
                        example: teacher@institute.local
                      new_email:
                        type: string
                        example: new.teacher@institute.local
                      email_verified_at:
                        type: string
                        example: null
                      phone:
                        type: string
                        example: null
                      is_active:
                        type: boolean
                        example: true
                      default_path:
                        type: string
                        example: /api/gradebooks
                      gradebooks_count:
                        type: integer
                        example: 2
                      created_at:
                        type: string
                        example: '2026-05-16T12:00:00+00:00'
                      updated_at:
                        type: string
                        example: '2026-05-16T12:15:00+00:00'
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Необходима авторизация.'
                properties:
                  message:
                    type: string
                    example: 'Необходима авторизация.'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Этот email уже используется.'
                  errors:
                    email:
                      - 'Этот email уже используется.'
                properties:
                  message:
                    type: string
                    example: 'Этот email уже используется.'
                  errors:
                    type: object
                    properties:
                      email:
                        type: array
                        example:
                          - 'Этот email уже используется.'
                        items:
                          type: string
      tags:
        - Общее
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Новый email.'
                  example: new.user@example.com
                password:
                  type: string
                  description: 'Текущий пароль.'
                  example: password
              required:
                - email
                - password
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Получите токен через <code>POST /api/auth/login</code> (email и пароль). Передавайте заголовок <code>Authorization: Bearer {TOKEN}</code>.'
security:
  -
    default: []